pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rl-robot"
version = "0.1.0"
description = "RL environment for wet shotcrete robot research."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "hydra-core>=1.3",
    "matplotlib>=3.8",
    "numpy>=2.0",
    "omegaconf>=2.3",
    "plotly>=5.0",
    "pyyaml>=6.0",
    "torch>=2.3",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4"]
noise = ["opensimplex>=0.4.5"]
visualization = ["pybullet>=3.2"]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["rl_robot*"]

[tool.setuptools.package-data]
rl_robot = [
    "py.typed",
    "assets/**/*.yaml",
    "assets/**/*.json",
    "assets/**/*.urdf",
    "assets/**/*.obj",
    "assets/**/*.html",
]

[tool.uv]
package = true
